ajoute la possibilité de générer un sommaire dans les docs mdx [ci skip]
