+++ site/app/cgu/cgu.mdx
@@ -1,4 +1,4 @@
-### 1. Préambule
+## Préambule
 
 1.1. Les présentes Conditions Générales d’Utilisation (ci-après les « CGU ») décrivent les termes et conditions dans lesquels l'ADEME fournit la plateforme Territoires en Transitions à l’utilisateur (ci-après « utilisateur »), accessible à l’adresse : [https://territoiresentransitions.fr/](https://territoiresentransitions.fr/)
 
@@ -16,13 +16,13 @@
 
 1.8. Le contenu de la plateforme n’est pas que de la responsabilité de l’ADEME. L’utilisateur accepte et reconnaît qu’il est seul responsable des informations, textes, données, fichiers etc. qu’il publie.
 
-### 2. Contrat conclu sous forme électronique
+## Contrat conclu sous forme électronique
 
 2.1. L’utilisateur reconnaît être parfaitement informé du fait que son accord concernant le contenu des présentes CGU ne nécessite pas la signature manuscrite de ce document, mais résulte de sa seule acceptation en ligne.
 
 2.2. En conséquence, il est expressément convenu que l’acceptation en ligne des présentes CGU en constitue une acceptation formelle et lie contractuellement les deux parties.
 
-### 3. Définitions
+## Définitions
 
 Les termes ci-dessous définis auront entre les parties la signification suivante :
 
@@ -41,13 +41,13 @@ Les termes ci-dessous définis auront entre les parties la signification suivant
   4. « Visite » : accès en lecture à une partie du profil de la collectivité.
 - « Identité » : Nom et prénom de l’utilisateur permettant de l’identifier eu égard aux opérations qu’il effectue sur la plateforme et de le contacter dans le cadre des échanges de bonnes pratiques entre collectivités et utilisateurs de la plateforme.
 
-### 4. Objet
+## Objet
 
 Les présentes CGU visent à définir les CGU de la plateforme accessible à l’adresse [https://territoiresentransitions.fr/](https://territoiresentransitions.fr/).
 
 Dans le cas où l'Utilisateur ne souhaite pas accepter tout ou partie des présentes conditions générales, il lui est demandé de renoncer à tout usage du Service.
 
-### 5. Entrée en vigueur – Durée
+## Entrée en vigueur – Durée
 
 5.1. L’utilisation du site [https://territoiresentransitions.fr/](https://territoiresentransitions.fr/) implique l’acceptation pleine et entière des conditions générales d’utilisation ci-après décrites.
 
@@ -55,7 +55,7 @@ Dans le cas où l'Utilisateur ne souhaite pas accepter tout ou partie des prése
 
 5.3. Ces conditions d’utilisation sont susceptibles d’être modifiées ou complétées à tout moment, les utilisateurs du site [https://territoiresentransitions.fr/](https://territoiresentransitions.fr/) sont donc invités à les consulter de manière régulière.
 
-### 6. Modifications des CGU
+## Modifications des CGU
 
 6.1. L’équipe Territoires en Transitions et l’ADEME se réservent le droit de modifier à tout moment les présentes CGU.
 
@@ -63,21 +63,21 @@ Dans le cas où l'Utilisateur ne souhaite pas accepter tout ou partie des prése
 
 6.3. Les CGU figurant en ligne sont accessibles depuis le pied de page, et prévalent sur toute version imprimée de date antérieure.
 
-### 7. Résiliation
+## Résiliation
 
 L’ADEME se réserve le droit de supprimer définitivement, en tout ou partie, la plateforme Territoires en Transitions et de mettre fin à tout moment à l’utilisation de la plateforme par l’utilisateur.
 
-### 8. Gratuité du service
+## Gratuité du service
 
 La plateforme Territoires en Transitions est mise à disposition gratuitement.
 
 L’ensemble des frais d’accès à la plateforme reste à la charge de l’utilisateur.
 
-### 9. Suspension de l’accès à la plateforme
+## Suspension de l’accès à la plateforme
 
 L’équipe Territoires en Transitions et l’ADEME se réserve le droit, sans préavis, de suspendre temporairement l’accès à tout ou partie de la plateforme pour des raisons techniques liées notamment à la nécessité de mettre à jour la plateforme, de la modifier, d’en assurer la maintenance et, de manière générale, pour tout autre cause technique et/ou organisationnelle et/ou de sécurité.
 
-### 10. Obligations des utilisateurs
+## Obligations des utilisateurs
 
 10.1. L’utilisateur s’engage à disposer de tous les moyens nécessaires visant à protéger son poste de travail informatique contre toute attaque de virus informatique. L'utilisateur sera tenu responsable de tout dommage résultant du non-respect des règles de l'art en terme de protection virale.
 
@@ -98,7 +98,7 @@ L’équipe Territoires en Transitions et l’ADEME se réserve le droit, sans p
 
 10.4. L’équipe Territoires en Transitions se réserve également le droit de suspendre, limiter ou supprimer l’accès à la plateforme de l’utilisateur fautif.
 
-### 11. Responsabilités
+## Responsabilités
 
 11.1. L’ADEME ne saurait être tenue responsable de tout contenu (contraire aux obligations telles que définies dans les présentes) diffusé sur la plateforme Territoires en Transitions par les utilisateurs qui en sont seuls responsables.
 
@@ -113,7 +113,7 @@ L’équipe Territoires en Transitions et l’ADEME se réserve le droit, sans p
 
 11.5. L’ADEME décline, en conséquence, toute responsabilité dans le cas où le contenu de ces sites contreviendrait aux dispositions légales et réglementaires en vigueur.
 
-### 12. Gestion des identifiants
+## Gestion des identifiants
 
 12.1. Pour accéder à la plateforme Territoires en Transitions, l’utilisateur peut être amené à disposer d’un identifiant et d’un mot de passe (ceux demandés lors de la création de compte) pour s’authentifier.
 
@@ -127,7 +127,7 @@ L’équipe Territoires en Transitions et l’ADEME se réserve le droit, sans p
 
 12.6. L’utilisateur s’engage à ne pas s’approprier l’identifiant d’un autre utilisateur.
 
-### 13. Propriété intellectuelle
+## Propriété intellectuelle
 
 ### Propriété intellectuelle de l’ADEME
 
@@ -149,13 +149,13 @@ L’équipe Territoires en Transitions et l’ADEME se réserve le droit, sans p
 - du droit d’adapter, de faire évoluer, de réaliser de nouvelles versions ou de nouveaux développements des contenus, de les maintenir, modifier, assembler, transcrire, arranger, numériser, monter, amputer, condenser, migrer et étendre ;
 - du droit de représenter, de diffuser ou de faire diffuser les contenus par tous moyens et/ou supports électroniques, numériques, informatiques, de télécommunication connus ou inconnus à ce jour, et ce auprès de tout public.
 
-### 14. Confidentialité
+## Confidentialité
 
 14.1. L’utilisateur s’engage à respecter l’obligation de secret et de confidentialité la plus absolue sur les informations confidentielle dont il a eu connaissance au sein de la plateforme.
 
 14.2. A ce titre, l’utilisateur s’engage à ne permettre à aucun tiers l’accès à la plateforme, et à ne communiquer à aucun tiers quelconque de ces informations confidentielles.
 
-### 15. Données à caractère personnel
+## Données à caractère personnel
 
 15.1. L’ADEME pourra collecter des données personnelles concernant les utilisateurs dans le cadre de l’accès et de l’utilisation de la plateforme. Les informations personnelles demandées à chaque utilisateur sont limitées au strict nécessaire :
 
@@ -177,7 +177,7 @@ Délégué à la Protection des Données/Data Protection Officer
 
 15.5. L’ADEME est attachée au respect des règles de protection de la vie privée des utilisateurs de ses sites internet, de ses espaces collaboratifs et de ses applications mobiles. Nous nous engageons à n’utiliser les informations communiquées lors de votre inscription à ce site que pour les traitements nécessaires à son fonctionnement et pour réaliser ponctuellement des enquêtes auprès des utilisateurs, nous permettant d’améliorer la pertinence et la qualité des informations traitées et diffusées. [Consultez la politique de protection des données personnelles de l'ADEME](https://www.ademe.fr/donnees-personnelles/)
 
-### 16. Droit applicable en cas de litige
+## Droit applicable en cas de litige
 
 16.1. Les présentes CGU sont régies par le droit français.
