+++ site/app/cgu/page.tsx
@@ -1,8 +1,8 @@
-'use client';
-
-import CGU from './cgu.mdx';
+import dynamic from 'next/dynamic';
 
 export default function Page() {
+  const CGU = dynamic(import('./cgu.mdx'));
+
   return (
     <div className="fr-container">
       <div className="fr-mt-1w fr-mt-md-4w fr-mb-5w">
