+++ site/app/global.css
@@ -1,0 +1,29 @@
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
+
+html,
+body {
+  max-width: 100vw;
+  overflow-x: hidden;
+}
+
+/** appliquable au sommaire généré à partir d'un mdx avec rehype-toc  */
+nav.md-toc {
+  background-color: #f6f6f6;
+  padding: 1rem 1.5rem;
+  margin-bottom: 2rem;
+}
+
+/** affiche la numérotation des h2 correspondant aux entrées du sommaire  */
+/** ref: https://gist.github.com/patik/89ee6092c72a9e39950445c01598517a */
+body,
+h1 {
+  counter-reset: h2counter;
+}
+nav.md-toc + h2:before {
+  counter-increment: h2counter;
+  content: counter(h2counter) '.\0000a0\0000a0';
+}
