+++ site/package.json
@@ -33,6 +33,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-leaflet": "^4.2.0",
+    "rehype-slug": "^5.1.0",
+    "rehype-toc": "^3.0.2",
     "swr": "^2.0.0",
     "typescript": "4.9.4"
   }
